add approximated location
Besides the real position of the agent, the agent can only receive a location input that is distorted by noise, and the noise will increase with the distance from the starting point. When agent is close to the goal, which can be set manually, the noise will be 0, as we assume now agent can see the goal clearly without using approximation.